Rename MessageBox to MessageBoxWindow and extract Show() methods to separate MessageBox class
