+++ AdonisUI/Themes/Generic.xaml
@@ -5,7 +5,7 @@
     <ResourceDictionary.MergedDictionaries>
 
         <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/AdonisWindow.xaml"/>
-        <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/MessageBox.xaml"/>
+        <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/MessageBoxWindow.xaml"/>
         <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/RippleHost.xaml"/>
         <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/SplitButton.xaml"/>
         <ResourceDictionary Source="pack://application:,,,/AdonisUI;component/Controls/ValidationErrorIndicator.xaml"/>
